스킬 시스템 확장 및 디버프 구조 개선, 타게팅 시스템 고도화
### 1. 스킬 시스템 확장
- 즉발형 AoE 스킬 처리 로직 추가
- 멀티 타겟 스킬 구조 정비 (여러 명 타격 / 범위 공격 기반 준비)
- 데미지 계산이 여러 대상에게 적용되도록 SkillResolver 확장

### 2. 타게팅 시스템 고도화
- SkillTargetingType 종류 대폭 확장
  - 가장 가까운 대상 / 가장 먼 대상
  - HP 가장 낮은 대상 / ATK 가장 높은 대상
  - 랜덤 대상
  - 가까운 N명 / 먼 N명
  - 적 전체 / 아군 전체
  - (추가기반) 원형 / 부채꼴 / 직선 범위 타게팅 구조 설계
- TargetResolver에 타게팅 타입별 분기 처리 추가
- SelectNearest / SelectFarthest 등 N명 선택 함수 구현

### 3. 디버프 및 CC 시스템 개선
- Stun / Silence / Freeze / Root 지속시간 감소 적용
- DebuffTickSystem에서 지속시간 Tick 처리 및 해제 처리 구현
- Knockback / Pull / Push 구조 정비

### 4. 버프/디버프 스탯 반영 구조 개선
- BuffStatSystem에서 BuffKind별 스탯 반영 방식 정리
- Effective Stat 초기화 → 누적 반영 구조 유지
- 공격력/방어력/속도/크리율/크리데미지 등 안전한 클램핑 처리 적용

### 5. SkillDebuffSystem 호출 구조 수정
- Apply() 시 hitIndex, extraMultiplier 누락 문제 해결
- SkillResolver에서 Debuff 적용 시 인자 전달 방식 정리

### 6. 데이터 테이블 작업 (SQL)
- 스킬 레벨별 Values JSON 정리
- 패시브/힐/버프/디버프/부가 효과 등 모두 표준화된 구조로 저장

---

전체적으로 오늘 작업은
- “스킬이 누구에게 영향을 미치는지”
- “어떤 방식으로 효과가 적용되는지”
- “버프/디버프/CC가 어떻게 누적되는지”

+++ Client/Assets/Script/Network/GamePlay/CombatNetwork.cs
@@ -55,7 +55,7 @@ public class CombatNetwork
     public void SendCommand(
     long combatId,
     long actorId,
-    long skillId,
+    int skillId,
     long? targetActorId = null,
     Action<bool> onDone = null)
     {
